requirements.txt
absl-py==1.4.0
astunparse==1.6.3
backports.zoneinfo==0.2.1
bsuite==0.3.5
cachetools==5.3.0
certifi==2022.12.7
chai-sacred==0.8.3
charset-normalizer==3.1.0
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dm-env==1.6
dm-sonnet==2.0.1
dm-tree==0.1.8
docopt==0.6.2
filelock==3.10.0
flatbuffers==2.0.7
fonttools==4.39.0
frozendict==2.3.5
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gym==0.21.0
gym-notices==0.0.8
gymnasium==0.27.1
gymnasium-notices==0.0.1
h5py==3.8.0
huggingface-hub==0.13.3
huggingface-sb3==2.2.4
idna==3.4
imageio==2.26.0
imitation==0.3.2
importlib-metadata==4.13.0
importlib-resources==5.12.0
jax-jumpy==0.2.0
joblib==1.2.0
jsonpickle==3.0.1
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.1
libclang==15.0.6.1
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mizani==0.8.1
munch==2.5.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
palettable==3.3.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
plotnine==0.10.1
protobuf==3.19.6
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.27
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seals==0.1.5
six==1.16.0
smmap==5.0.0
stable-baselines3==1.7.0
statsmodels==0.13.5
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.19.0
termcolor==2.2.0
threadpoolctl==3.1.0
tianshou==0.5.0
tifffile==2023.2.28
torch==1.13.1
tqdm==4.65.0
trfl==1.2.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
